fix(tests): fix narrowing conversion and deprecation warnings
- Fix narrowing conversion error in test_sparse_vector_neighbor_exchange.cpp
  by explicitly casting int to size_t
- Replace all deprecated world::create(Int3, Real3, Real3) calls with
  world::create(GridSize, PhysicalOrigin, GridSpacing) for type safety
- Add strong_types.hpp include to all test files using the new API
- Fix incorrectly converted single-parameter calls in test_world.cpp
- Format all modified test files with clang-format

+++ tests/unit/core/test_sparse_vector_neighbor_exchange.cpp
@@ -97,7 +97,9 @@ TEST_CASE("Neighbor exchange - Ring topology",
   int left_neighbor = (rank - 1 + size) % size;
 
   // Create data to send
-  std::vector<size_t> send_indices = {rank * 10, rank * 10 + 1, rank * 10 + 2};
+  std::vector<size_t> send_indices = {static_cast<size_t>(rank * 10),
+                                      static_cast<size_t>(rank * 10 + 1),
+                                      static_cast<size_t>(rank * 10 + 2)};
   std::vector<double> send_data = {static_cast<double>(rank * 100),
                                    static_cast<double>(rank * 100 + 10),
                                    static_cast<double>(rank * 100 + 20)};
